Removed old CORDIC.py. Renamed CORDIC_v2.py --> CORDIC.py
